+++ AdventOfCode/2019/Day09.cs
@@ -10,7 +10,7 @@ namespace Lomont.AdventOfCode._2019
             var lines = ReadLines();
             var prog = Numbers64(lines[0]);
             List<long> output = new();
-            var mem = Day02.RunIntCode(prog, new List<int> { part2?2:1}, s=>output.Add(s));
+            var mem = Day02.RunIntCode(prog, new List<long> { part2?2:1}, s=>output.Add(s));
             return output.Last();
         }
     }
